% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleApp.R
\name{alleleApp}
\alias{alleleApp}
\alias{alleleServer}
\alias{alleleInput}
\alias{alleleUI}
\alias{alleleOutput}
\title{Shiny Allele App}
\usage{
alleleApp()

alleleServer(
  id,
  hotspot_list,
  pattern_list,
  pairprobs_obj,
  patterns,
  project_df,
  snp_action
)

alleleInput(id)

alleleUI(id)

alleleOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pattern_list, pairprobs_obj, patterns, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Allele App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
