% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleServer.R
\name{alleleServer}
\alias{alleleServer}
\alias{alleleUI}
\alias{alleleOutput}
\title{Shiny allele coefficient analysis and plot module}
\usage{
alleleServer(
  id,
  win_par,
  phe_mx,
  cov_df,
  probs_obj,
  K_chr,
  analyses_df,
  patterns,
  scan_pat,
  project_df,
  snp_action
)

alleleUI(id)

alleleOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, phe_mx, cov_df, probs_obj, K_chr, analyses_df, patterns, scan_pat, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 coefficient plots, with interfaces \code{alleleUI} and  \code{alleleOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
