% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collapse_covar}
\alias{collapse_covar}
\title{Collapse covariate names in analyses table}
\usage{
collapse_covar(analyses_tbl)
}
\arguments{
\item{analyses_tbl}{analyses table}
}
\description{
Collapse covariates used to comma-separated list and return table
}
\examples{
\dontrun{collapse_covar(analyses_tbl)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

