% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarApp.R
\name{covarApp}
\alias{covarApp}
\alias{covarServer}
\alias{covarOutput}
\title{Shiny Covariate App}
\usage{
covarApp()

covarServer(id, pheno_mx, project_df)

covarOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{pheno_mx, project_df}{reactive arguments}
}
\description{
Shiny Covariate App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
