% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrna_query_func.R
\name{create_mrna_query_func}
\alias{create_mrna_query_func}
\title{Create a function to query mRNA data}
\usage{
create_mrna_query_func(dbfile, mrnadir_val = "RNAseq")
}
\arguments{
\item{dbfile}{Name of database file}

\item{mrnadir_val}{name of mRNA data directory (default \code{"RNAseq"})}
}
\value{
Function with seven arguments, `chr`, `start`,
    `end`, `local`, `qtl`, `fast` and `mrnadir`. It returns a list with `expr`, `annot` and `peaks` objects
    spanning the region specified by the first three arguments.
}
\description{
Create a function that will connect to a database of mRNA information
and return a list with `probs` object and a `map` object.
}
\details{
Note that this function assumes positions are in Mbp.
    There are required columns for each element, to be detailed in time.
    See \code{\link{read_mrna}} for details on how mRNA data are read.
    See \code{\link[qtl2]{create_variant_query_func}} for original idea.
}
