% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diploApp.R
\name{diploApp}
\alias{diploApp}
\alias{diploServer}
\alias{diploUI}
\alias{diploOutput}
\title{Shiny Diplotype module}
\usage{
diploApp()

diploServer(id, hotspot_list, project_df)

diploUI(id)

diploOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, phe_mx, cov_df, K_chr, analyses_df, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny diplotype SNP/Gene action analysis, with interface \code{diploUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
