% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diploServer.R
\name{diploServer}
\alias{diploServer}
\alias{diploUI}
\title{Shiny Diplotype module}
\usage{
diploServer(
  id,
  win_par,
  phe_mx,
  cov_df,
  K_chr,
  analyses_df,
  project_df,
  allele_info
)

diploUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, phe_mx, cov_df, K_chr, analyses_df, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny diplotype SNP/Gene action analysis, with interface \code{diploUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
