% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipParApp.R
\name{dipParApp}
\alias{dipParApp}
\alias{dipParServer}
\alias{dipParInput}
\alias{dipParUI}
\alias{dipParOutput}
\title{Shiny Diplotype Parameter module}
\usage{
dipParApp()

dipParServer(id)

dipParInput(id)

dipParUI(id)

dipParOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, phe_mx, cov_df, K_chr, analyses_df, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Diplotype Parameter module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
