% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQtl2.R
\name{fitQtl2}
\alias{fitQtl2}
\title{Wrapper for 'qtl2' function fit1}
\usage{
fitQtl2(driver, target, ...)
}
\arguments{
\item{driver}{genotype probabilities in matrix}

\item{target}{numeric vector of target values}

\item{...}{additional parameters for \code{\link[qtl2]{fit1}}}
}
\description{
Wrapper for 'qtl2' function fit1
}
