% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneRegionApp.R
\name{geneRegionApp}
\alias{geneRegionApp}
\alias{geneRegionServer}
\alias{geneRegionInput}
\alias{geneRegionOutput}
\title{Shiny Genes in SNP Region module}
\usage{
geneRegionApp()

geneRegionServer(id, snp_list, project_df)

geneRegionInput(id)

geneRegionOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{geneRegionInput}, \code{geneRegionUI} and  \code{geneRegionOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
