% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoApp.R
\name{genoApp}
\alias{genoApp}
\alias{genoServer}
\alias{genoInput}
\alias{genoOutput}
\title{Shiny Allele App}
\usage{
genoApp()

genoServer(id, hotspot_list, snp_list, pairprobs_obj, project_df)

genoInput(id)

genoOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pairprobs_obj, patterns, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Allele App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
