% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDataApp.R
\name{genoDataApp}
\alias{genoDataApp}
\alias{genoDataServer}
\alias{genoDataInput}
\alias{genoDataOutput}
\title{Shiny Genotype Data App}
\usage{
genoDataApp()

genoDataServer(id, hotspot_list, snp_list, pairprobs_obj, project_df)

genoDataInput(id)

genoDataOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pairprobs_obj, patterns, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Genotype Data App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
