% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoEffectApp.R
\name{genoEffectApp}
\alias{genoEffectApp}
\alias{genoEffectServer}
\alias{genoEffectUI}
\alias{genoEffectOutput}
\title{Shiny Allele App}
\usage{
genoEffectApp()

genoEffectServer(
  id,
  hotspot_list,
  pattern_list,
  snp_list,
  geno_list,
  pairprobs_obj,
  project_df
)

genoEffectUI(id)

genoEffectOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pattern_list, pairprobs_obj, patterns, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Allele App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
