% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPanelApp.R
\name{genoPanelApp}
\alias{genoPanelApp}
\alias{genoPanelServer}
\alias{genoPanelInput}
\alias{genoPanelOutput}
\title{Shiny Allele App}
\usage{
genoPanelApp()

genoPanelServer(
  id,
  hotspot_list,
  pattern_list,
  snp_list,
  pairprobs_obj,
  project_df
)

genoPanelInput(id)

genoPanelOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pattern_list, pairprobs_obj, patterns, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Allele App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
