% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotApp.R
\name{genoPlotApp}
\alias{genoPlotApp}
\alias{genoPlotServer}
\alias{genoPlotInput}
\alias{genoPlotOutput}
\title{Shiny Genotype Data App}
\usage{
genoPlotApp()

genoPlotServer(id, hotspot_list, geno_list)

genoPlotInput(id)

genoPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pairprobs_obj, patterns, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Genotype Data App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
