% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covar.R
\name{get_covar}
\alias{get_covar}
\title{Get covariates}
\usage{
get_covar(covar, analyses_tbl)
}
\arguments{
\item{covar}{matrix of all covariates}

\item{analyses_tbl}{table of analyses setups}
}
\value{
matrix of covariates
}
\description{
Get covariates using \code{analyses_tbl} filtered by \code{phename_output}.
}
\examples{
dirpath <- "https://raw.githubusercontent.com/rqtl/qtl2data/master/DOex"

# Read DOex example covar from 'qtl2data'
covar <- read.csv(file.path(dirpath, "DOex_covar.csv"))

# Select Sex and Cohort columns of covariates
analyses_tbl <- data.frame(Sex = TRUE, Cohort = TRUE)
# Pull covariates
covar <- get_covar(covar, analyses_tbl)
tidyr::pivot_wider(
  dplyr::count(covar, Sex, Cohort),
  names_from = "Sex", values_from = "n")

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
