% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Get gene in region}
\usage{
get_genes(chr_id, start, stop, gene_tbl = query_genes(chr_id, start, stop))
}
\arguments{
\item{chr_id}{chromosome identifier}

\item{start}{start position in Mbp}

\item{stop}{stop position in Mbp}

\item{gene_tbl}{table of genes from user supplied \code{query_genes}; see \code{\link[qtl2]{create_gene_query_func}}}
}
\value{
object of class \code{feature_tbl}.
}
\description{
Get gene using \code{query_genes}; see \code{\link[qtl2]{create_gene_query_func}}.
}
