% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploApp.R
\name{haploApp}
\alias{haploApp}
\alias{haploServer}
\alias{haploUI}
\title{Shiny Haplotype Analysis App}
\usage{
haploApp()

haploServer(
  id,
  win_par,
  pmap_obj,
  phe_mx,
  cov_df,
  K_chr,
  analyses_df,
  covar,
  analyses_tbl,
  peak_df,
  project_df,
  allele_info
)

haploUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, pmap_obj, phe_mx, cov_df, K_chr, analyses_df, covar, analyses_tbl, peak_df, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for analysis based on haplotype alleles, with interface \code{haploUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
