% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapParApp.R
\name{hapParApp}
\alias{hapParApp}
\alias{hapParServer}
\alias{hapParInput}
\alias{hapParUI}
\alias{hapParOutput}
\title{Shiny Haplotype Analysis App}
\usage{
hapParApp()

hapParServer(id)

hapParInput(id)

hapParUI(id)

hapParOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, pmap_obj, pheno_mx, covar_df, K_chr, analyses_df, covar, analyses_tbl, peak_df, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for analysis based on haplotype alleles, with interface \code{haploUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
