% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpPopup.R
\name{helpPopup}
\alias{helpPopup}
\title{Popup shiny function from Joe Cheng}
\usage{
helpPopup(
  title,
  content,
  placement = c("right", "top", "left", "bottom"),
  trigger = c("click", "hover", "focus", "manual")
)
}
\arguments{
\item{title}{Title of help popup}

\item{content}{Content of help as character string}

\item{placement}{Placement position}

\item{trigger}{Trigger mechanism}
}
\description{
Popup shiny function from Joe Cheng
}
\details{
{
See <https://gist.github.com/jcheng5/5913297>.
}
}
