% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotAccordion.R
\name{hotspotAccordionApp}
\alias{hotspotAccordionApp}
\title{Shiny Hotspot Panel App}
\usage{
hotspotAccordionApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Hotspot Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
