% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotApp.R
\name{hotspotApp}
\alias{hotspotApp}
\alias{hotspotServer}
\alias{hotspotInput}
\alias{hotspotUI}
\alias{hotspotOutput}
\title{Shiny Hotspot Panel App}
\usage{
hotspotApp()

hotspotServer(id, project_df, active_panel = shiny::reactive(TRUE))

hotspotInput(id)

hotspotUI(id)

hotspotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Hotspot Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
