% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotDataApp.R
\name{hotspotDataApp}
\alias{hotspotDataApp}
\alias{hotspotDataServer}
\alias{hotspotDataInput}
\alias{hotspotDataOutput}
\title{Shiny Hotspot Data App}
\usage{
hotspotDataApp()

hotspotDataServer(id, set_par, peak_df, pmap_obj, project_df)

hotspotDataInput(id)

hotspotDataOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\value{
list of inputs and scan summary

No return value; called for side effects.
}
\description{
Shiny module to select hotspots for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
