% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotPanelApp.R
\name{hotspotPanelApp}
\alias{hotspotPanelApp}
\alias{hotspotPanelServer}
\alias{hotspotPanelInput}
\alias{hotspotPanelUI}
\alias{hotspotPanelOutput}
\alias{hotspotPhenoOutput}
\alias{hotspotHotspotOutput}
\title{Shiny Hotspot Panel App}
\usage{
hotspotPanelApp()

hotspotPanelServer(id, project_df)

hotspotPanelInput(id)

hotspotPanelUI(id)

hotspotPanelOutput(id)

hotspotPhenoOutput(id)

hotspotHotspotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Hotspot Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
