% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotPlotApp.R
\name{hotspotPlotApp}
\alias{hotspotPlotApp}
\alias{hotspotPlotServer}
\alias{hotspotPlotOutput}
\title{Shiny Hotspot Plot App}
\usage{
hotspotPlotApp()

hotspotPlotServer(id, set_par, hotspot_obj)

hotspotPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, hotspot_obj}{reactive arguments}
}
\value{
list of inputs and scan summary

No return value; called for side effects.
}
\description{
Shiny module to view hotspots for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
