% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotTableApp.R
\name{hotspotTableApp}
\alias{hotspotTableApp}
\alias{hotspotTableServer}
\alias{hotspotTableOutput}
\title{Shiny Hotspot Table App}
\usage{
hotspotTableApp()

hotspotTableServer(id, hotspot_obj)

hotspotTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_obj}{reactive arguments}
}
\value{
list of inputs and scan summary

No return value; called for side effects.
}
\description{
Shiny module to select hotspots for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
