% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipApp.R
\name{kinshipApp}
\alias{kinshipApp}
\alias{kinshipServer}
\alias{kinshipOutput}
\title{Shiny Kinship App}
\usage{
kinshipApp()

kinshipServer(id, win_par, project_df)

kinshipOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, project_df}{reactive arguments}
}
\description{
Shiny Kinship App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
