% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainApp.R
\name{mainApp}
\alias{mainApp}
\alias{mainServer}
\alias{mainInput}
\alias{mainOutput}
\title{Shiny app for qtl2}
\usage{
mainApp()

mainServer(id, projects_df)

mainInput(id)

mainOutput(id)
}
\arguments{
\item{id}{identifier for shiny module}

\item{projects_df}{static data frame with project information}
}
\value{
No return value; called for side effects.
}
\description{
Run shiny app for qtl2 data.
}
\details{
See qtl2shinyData vignette for data setup.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
