% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainServer.R
\name{mainServer}
\alias{mainServer}
\alias{mainInput}
\alias{mainOutput}
\alias{mainApp}
\title{Shiny app for qtl2}
\usage{
mainServer(id, projects)

mainInput(id)

mainOutput(id)

mainApp()
}
\arguments{
\item{id}{identifier for shiny module}

\item{projects}{static data frame with project information}
}
\value{
No return value; called for side effects.
}
\description{
Run shiny app for qtl2 data.
}
\details{
See qtl2shinyData vignette for data setup.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
