% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediateDataApp.R
\name{mediateDataApp}
\alias{mediateDataApp}
\alias{mediateDataServer}
\alias{mediateDataInput}
\alias{mediateDataUI}
\alias{mediateDataOutput}
\title{Shiny Mediate Data App}
\usage{
mediateDataApp()

mediateDataServer(id, hotspot_list, snp_list, probs_obj, project_df)

mediateDataInput(id)

mediateDataUI(id)

mediateDataOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, patterns, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Mediate Data App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
