% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatePanelApp.R
\name{mediatePanelApp}
\alias{mediatePanelApp}
\alias{mediatePanelServer}
\alias{mediatePanelInput}
\alias{mediatePanelOutput}
\title{Shiny Haplotype Analysis App}
\usage{
mediatePanelApp()

mediatePanelServer(id, hotspot_list, snp_list, probs_obj, project_df)

mediatePanelInput(id)

mediatePanelOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, snp_list, probs_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for analysis based on haplotype alleles, with interface \code{mediatePanelUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
