% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatePlotApp.R
\name{mediatePlotApp}
\alias{mediatePlotApp}
\alias{mediatePlotServer}
\alias{mediatePlotInput}
\alias{mediatePlotOutput}
\title{Shiny Mediate Plot App}
\usage{
mediatePlotApp()

mediatePlotServer(id, hotspot_list, snp_list, probs_obj, project_df)

mediatePlotInput(id)

mediatePlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, patterns, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Mediate Plot App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
