% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medListApp.R
\name{medListApp}
\alias{medListApp}
\alias{medListServer}
\alias{medListInput}
\alias{medListOutput}
\title{Shiny coefficient analysis and plot module}
\usage{
medListApp()

medListServer(id, hotspot_list, snp_list, probs_obj, project_df)

medListInput(id)

medListOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, patterns, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny coefficient analysis and plot module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
