% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_pheno.R
\name{num_pheno}
\alias{num_pheno}
\title{Number of phenotypes examined}
\usage{
num_pheno(pheno, analyses_tbl)
}
\arguments{
\item{pheno}{names of phenotypes}

\item{analyses_tbl}{table of analyses run}
}
\description{
Character string with number of phenotypes of total.
}
\examples{
\dontrun{num_pheno(pheno, analyses_tbl)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

