% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternApp.R
\name{patternApp}
\alias{patternApp}
\alias{patternServer}
\alias{patternInput}
\alias{patternOutput}
\title{Shiny Pattern Panel App}
\usage{
patternApp()

patternServer(id, dip_par, hotspot_list, snp_list, pairprobs_obj, project_df)

patternInput(id)

patternOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{dip_par, hotspot_list, snp_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Pattern Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
