% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternDataApp.R
\name{patternDataApp}
\alias{patternDataApp}
\alias{patternDataServer}
\alias{patternDataInput}
\alias{patternDataUI}
\alias{patternDataOutput}
\title{Shiny Pattern Data App}
\usage{
patternDataApp()

patternDataServer(
  id,
  hotspot_list,
  dip_par,
  pairprobs_obj,
  patterns,
  snp_action = shiny::reactive({
     "basic"
 }),
  project_df
)

patternDataInput(id)

patternDataUI(id)

patternDataOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, dip_par, pairprobs_obj, patterns, snp_action, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for SNP pattern plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
