% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternPanelApp.R
\name{patternPanelApp}
\alias{patternPanelApp}
\alias{patternPanelServer}
\alias{patternPanelInput}
\alias{patternPanelOutput}
\title{Shiny Pattern Panel App}
\usage{
patternPanelApp()

patternPanelServer(
  id,
  dip_par,
  hotspot_list,
  snp_list,
  pairprobs_obj,
  project_df
)

patternPanelInput(id)

patternPanelOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{dip_par, hotspot_list, snp_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Pattern Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
