% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternPlotApp.R
\name{patternPlotApp}
\alias{patternPlotApp}
\alias{patternPlotServer}
\alias{patternPlotOutput}
\title{Shiny Pattern Plot App}
\usage{
patternPlotApp()

patternPlotServer(id, pattern_list, pairprobs_obj)

patternPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, dip_par, pairprobs_obj, patterns, snp_action, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Pattern Plot App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
