% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternServer.R
\name{patternServer}
\alias{patternServer}
\alias{patternUI}
\alias{patternOutput}
\title{Shiny Pattern module}
\usage{
patternServer(
  id,
  job_par,
  chr_pos,
  win_par,
  phe_mx,
  cov_df,
  pairprobs_obj,
  K_chr,
  analyses_df,
  patterns,
  project_df,
  allele_info,
  snp_action = shiny::reactive({
     NULL
 })
)

patternUI(id)

patternOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{job_par, chr_pos, win_par, phe_mx, cov_df, pairprobs_obj, K_chr, analyses_df, patterns, project_df, allele_info, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for SNP pattern plots, with interfaces \code{patternUI} and  \code{patternOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
