% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_class.R
\name{peak_class}
\alias{peak_class}
\alias{read_peak_class}
\alias{create_peak_class}
\title{Read Peaks for Phenotypes by Class}
\usage{
peak_class(
  project_df,
  class_name = NULL,
  subjects = c("all_mice", "HC_mice", "HF_mice", "male_mice", "female_mice"),
  covars = c("additive", "diet_interactive", "sex_interactive"),
  ...
)

read_peak_class(project_df, class_name = NULL, ...)

create_peak_class(project_df, force = FALSE)
}
\arguments{
\item{project_df}{data frame with project information}

\item{class_name}{character string of phenotype class}

\item{subjects}{character string of subjects used in analysis}

\item{covars}{character string of covariate selection}

\item{...}{additional parameters not used}

\item{force}{force creation if `TRUE`}
}
\value{
dataframe of peaks

character list of created RDS files
}
\description{
Read Peaks for Phenotypes by Class

Create Peaks Directory with Phenotypes by Class
}
