% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakApp.R, R/peakPanelApp.R
\name{peakApp}
\alias{peakApp}
\alias{peakServer}
\alias{peakOutput}
\alias{peakPanelOutput}
\title{Shiny Peak App}
\usage{
peakApp()

peakServer(id, set_par, project_df)

peakOutput(id)

peakPanelOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
