% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFilterApp.R
\name{peakFilterApp}
\alias{peakFilterApp}
\alias{peakFilterServer}
\alias{peakFilterInput}
\title{Shiny Phenotype App}
\usage{
peakFilterApp()

peakFilterServer(id, set_par, win_par, peak_df, project_df)

peakFilterInput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
