% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPanelApp.R
\name{peakPanelApp}
\alias{peakPanelApp}
\alias{peakPanelServer}
\alias{peakPanelInput}
\title{Shiny Phenotype App}
\usage{
peakPanelApp()

peakPanelServer(id, set_par, win_par, peak_project_df, project_df)

peakPanelInput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_project_df, pmap_obj, project_df}{reactive arguments}
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
