% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakReadApp.R
\name{peakReadApp}
\alias{peakReadApp}
\alias{peakReadServer}
\alias{peakReadOutput}
\title{Shiny Peak Read App}
\usage{
peakReadApp()

peakReadServer(id, set_par, project_df)

peakReadOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
