% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksServer.R
\name{peaksServer}
\alias{peaksServer}
\alias{peaksInput}
\alias{peaksUI}
\alias{peaksOutput}
\title{Shiny peaks module}
\usage{
peaksServer(id, set_par, pheno_type, peaks_tbl, pmap_obj, project_info)

peaksInput(id)

peaksUI(id)

peaksOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, pheno_type, peaks_tbl, pmap_obj, project_info}{reactive arguments}
}
\value{
list of inputs and scan summary
}
\description{
Shiny module for peaks selection, with interfaces \code{peaksInput}, \code{peaksUI} and  \code{peaksOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
