% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_region.R
\name{pheno_region}
\alias{pheno_region}
\alias{get_expr_region}
\title{Create list with phenotypes in region}
\usage{
pheno_region(
  chr_id,
  start_val,
  end_val,
  covar,
  map,
  peaks,
  analyses,
  pheno_data,
  drivers = 2
)

get_expr_region(
  chr_id,
  start_val,
  end_val,
  covar,
  map,
  project_dir,
  drivers = 2,
  query_mrna = create_mrna_query_func(project_dir)
)
}
\arguments{
\item{chr_id, start_val, end_val}{chromosome and start and end value}

\item{covar}{covariate data frame}

\item{map}{list or vector of map positions}

\item{peaks}{table of peaks}

\item{analyses}{table of analyses}

\item{pheno_data}{matrix of phenotype data}

\item{drivers}{number of drivers (1 or 2; default is 2)}

\item{project_dir}{project directory with mRNA data in subdirector \code{RNAseq}}

\item{query_mrna}{query routine for mRNA data (see \code{\link{create_mrna_query_func}})}
}
\value{
list containing \code{pheno}, \code{annot} and \code{covar}.
}
\description{
Create list with phenotypes in region

Create list with expression phenotypes in region
}
\examples{
dirpath <- "https://raw.githubusercontent.com/rqtl/qtl2data/master/DOex"

# Read DOex example cross from 'qtl2data'
DOex <- qtl2::read_cross2(file.path(dirpath, "DOex.zip"))
DOex <- subset(DOex, chr = "2")

# Calculate genotype and allele probabilities
pr <- qtl2::calc_genoprob(DOex, error_prob=0.002)

# Summary of coefficients at scan peak
scan_pr <- qtl2::scan1(pr, DOex$pheno)
peaks <- summary(scan_pr, DOex$pmap)

# Select Sex and Cohort columns of covariates
analyses_tbl <- data.frame(pheno = "OF_immobile_pct", Sex = TRUE, Cohort = TRUE)

# Get phenos in region.
out <- pheno_region("2", 90, 100, DOex$covar, DOex$pmap, peaks, analyses_tbl, DOex$pheno)
str(out)

}
\seealso{
\code{\link{create_mrna_query_func}}
}
