% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoApp.R
\name{phenoApp}
\alias{phenoApp}
\alias{phenoServer}
\alias{phenoInput}
\alias{phenoUI}
\alias{phenoOutput}
\title{Shiny Hotspot Panel App}
\usage{
phenoApp()

phenoServer(id, set_par, win_par, peak_df, pmap_obj, hotspot_df, project_df)

phenoInput(id)

phenoUI(id)

phenoOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, win_par, peak_df, pmap_obj, hotspot_df, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Hotspot Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
