% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoDataApp.R
\name{phenoDataApp}
\alias{phenoDataApp}
\alias{phenoDataServer}
\alias{phenoDataInput}
\alias{phenoDataOutput}
\title{Shiny Phenotype Plot App}
\usage{
phenoDataApp()

phenoDataServer(id, pheno_names, pheno_mx, covar_df)

phenoDataInput(id)

phenoDataOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{pheno_names, pheno_mx, covar_df}{reactive arguments}
}
\value{
2-element vector of scan window
}
\description{
Shiny module to plot phenotypes.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
