% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoNamesApp.R
\name{phenoNamesApp}
\alias{phenoNamesApp}
\alias{phenoNamesServer}
\alias{phenoNamesInput}
\alias{phenoNamesOutput}
\title{Shiny Phenotype Names App}
\usage{
phenoNamesApp()

phenoNamesServer(id, set_par, peak_df, pheno_mx, covar_df, project_df)

phenoNamesInput(id)

phenoNamesOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, win_par, peak_df, analyses_df, covar, project_df}{reactive arguments}
}
\description{
Shiny module for phenotype name selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
