% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoPanelApp.R
\name{phenoPanelApp}
\alias{phenoPanelApp}
\alias{phenoPanelServer}
\alias{phenoPanelInput}
\alias{phenoPanelOutput}
\title{Shiny Hotspot Panel App}
\usage{
phenoPanelApp()

phenoPanelServer(
  id,
  set_par,
  win_par,
  peak_df,
  pmap_obj,
  hotspot_df,
  project_df
)

phenoPanelInput(id)

phenoPanelOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, win_par, peak_df, pmap_obj, hotspot_df, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Hotspot Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
