% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoPlotApp.R
\name{phenoPlotApp}
\alias{phenoPlotApp}
\alias{phenoPlotServer}
\alias{phenoPlotOutput}
\title{Shiny Phenotype Plot App}
\usage{
phenoPlotApp()

phenoPlotServer(id, pheno_names, phe_mx, cov_df)

phenoPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{pheno_names, phe_mx, cov_df}{reactive arguments}
}
\value{
2-element vector of scan window
}
\description{
Shiny module to plot phenotypes.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
