% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoPlotServer.R
\name{phenoPlotServer}
\alias{phenoPlotServer}
\alias{phenoPlotUI}
\title{Shiny Phenotype Plot module}
\usage{
phenoPlotServer(id, pheno_names, phe_mx, cov_df)

phenoPlotUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{pheno_names, phe_mx, cov_df}{reactive arguments}
}
\value{
2-element vector of scan window
}
\description{
Shiny module to plot phenotypes, with interface \code{phenoPlotUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
