% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoServer.R
\name{phenoServer}
\alias{phenoServer}
\alias{phenoUI}
\alias{phenoOutput}
\title{Shiny phenotype selection}
\usage{
phenoServer(id, set_par, win_par, peaks_df, analyses_tbl, cov_df, project_info)

phenoUI(id)

phenoOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, win_par, peaks_df, analyses_tbl, cov_df, project_info}{reactive arguments}
}
\description{
Shiny module for phenotype selection, with interfaces \code{phenoUI} and  \code{phenoOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
