% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoTableApp.R
\name{phenoTableApp}
\alias{phenoTableApp}
\alias{phenoTableServer}
\alias{phenoTableOutput}
\title{Shiny Phenotype Table App}
\usage{
phenoTableApp()

phenoTableServer(id, pheno_mx, covar_df)

phenoTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{pheno_mx, covar_df}{reactive arguments}
}
\value{
2-element vector of scan window
}
\description{
Shiny module to plot phenotypes.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
