% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eff.R
\name{plot_eff}
\alias{plot_eff}
\title{Plot effects}
\usage{
plot_eff(pheno, scan_obj, eff_obj, xlim = NULL, main = pheno, ...)
}
\arguments{
\item{pheno}{name of phenotype for effect scan}

\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{eff_obj}{object of class \code{\link[doqtl2]{listof_scan1coefCC}}}

\item{xlim}{x limits for plot}

\item{main}{plot title}

\item{...}{parameters past to plot routine}
}
\description{
Plot of effects.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}

