% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_sex}
\alias{plot_sex}
\title{Plot across traits}
\usage{
plot_sex(phe, cov)
}
\arguments{
\item{phe}{data frame of phenotypes}

\item{cov}{matrix of covariates}
}
\description{
Plot of density and pairwise scatterplots.
}
\examples{
\dontrun{plot_sex(phe)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}

