% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmapApp.R
\name{pmapApp}
\alias{pmapApp}
\alias{pmapServer}
\alias{pmapOutput}
\title{Shiny Pmap App}
\usage{
pmapApp()

pmapServer(id, project_df)

pmapOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{project_df}{reactive arguments}
}
\description{
Shiny Pmap App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
