% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsApp.R
\name{probsApp}
\alias{probsApp}
\alias{probsServer}
\alias{pairProbsServer}
\alias{snpProbsServer}
\title{Shiny Probability modules}
\usage{
probsApp()

probsServer(id, win_par, project_df)

pairProbsServer(id, win_par, project_df)

snpProbsServer(id, win_par, pheno_names, project_df)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, pheno_names, project_df}{reactive arguments}
}
\value{
Object of class \code{probs}.
}
\description{
Shiny genotype probability access.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
