% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsServer.R
\name{probsServer}
\alias{probsServer}
\alias{pairProbsServer}
\alias{snpProbsServer}
\title{Shiny Probability modules}
\usage{
probsServer(id, win_par, project_info)

pairProbsServer(id, win_par, project_info)

snpProbsServer(id, win_par, pheno_names, project_info)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, pheno_names, project_info}{reactive arguments}
}
\value{
Object of class \code{probs}.
}
\description{
Shiny genotype probability access.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
