% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectApp.R
\name{projectApp}
\alias{projectApp}
\alias{projectServer}
\alias{projectUI}
\alias{projectOutput}
\title{Shiny Project App}
\usage{
projectApp()

projectServer(id, projects_df)

projectUI(id)

projectOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{projects_df}{static project info data frame}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for selection of project, with interface \code{projectUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
