% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectServer.R
\name{projectServer}
\alias{projectServer}
\alias{projectUI}
\title{Shiny project module}
\usage{
projectServer(id, projects_info)

projectUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{projects_info}{reactive project info data frame}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for selection of project, with interface \code{projectUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
