% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{qtl2shiny_plot_theme}
\alias{qtl2shiny_plot_theme}
\title{QTL2 Shiny Plot Theme}
\usage{
qtl2shiny_plot_theme(base_size = 9)
}
\arguments{
\item{base_size}{base font size (default 9)}
}
\value{
a ggplot2 theme
}
\description{
QTL2 Shiny Plot Theme
}
