% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{qtl2shiny_theme}
\alias{qtl2shiny_theme}
\title{QTL2 Shiny UI Theme}
\usage{
qtl2shiny_theme(font_size = "0.85rem")
}
\arguments{
\item{font_size}{base font size in rem (default "0.85rem")}
}
\value{
a \code{bslib::bs_theme} object
}
\description{
QTL2 Shiny UI Theme
}
