% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{qtl2shinyApp}
\alias{qtl2shinyApp}
\alias{shinyDash}
\title{Shiny app for qtl2}
\usage{
qtl2shinyApp()

shinyDash()
}
\value{
No return value; called for side effects.
}
\description{
Run shiny app for qtl2 data.
}
\details{
See qtl2shinyData vignette for data setup.
}
\examples{
\dontrun{qtl2shinyApp()}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
