% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrna.R
\name{read_mrna}
\alias{read_mrna}
\title{Read mrna expression object from file}
\usage{
read_mrna(
  chr_id = NULL,
  start_val = NULL,
  end_val = NULL,
  datapath,
  local = TRUE,
  qtl = FALSE,
  mrnadir = "RNAseq"
)
}
\arguments{
\item{chr_id}{vector of chromosome identifiers}

\item{start_val, end_val}{start and end values in Mbp}

\item{datapath}{name of folder with Derived Data}

\item{local}{read only mRNA values local to region if \code{TRUE};
otherwise include distal mRNA values that map to region}

\item{qtl}{read only mRNA values with QTL peak in region if \code{TRUE}}

\item{mrnadir}{name of directory with mRNA data}
}
\value{
list with \code{expr} = matrix of expression mRNA values in region and \code{annot} = data frame of annotations for mRNA.
}
\description{
Uses fst to read mrna expression object and associated annotations.
}
\details{
Reads `expr`, `peaks` and `annot` information on mRNA and combines into a list.
    The `expr` and `peaks` elements are stored in `fst` database,
    while `peaks` is an `RDS` database.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
