% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peaks.R
\name{read_peaks}
\alias{read_peaks}
\alias{create_peak_class}
\title{Read Peaks for Phenotypes by Class}
\usage{
read_peaks(project_df, class = NULL, legacy = FALSE, ...)

create_peak_class(project_df, force = FALSE)
}
\arguments{
\item{project_df}{data frame with project information}

\item{class}{name of class}

\item{legacy}{use legacy data if `TRUE`}

\item{...}{additional arguments}

\item{force}{force creation if `TRUE`}
}
\value{
dataframe of peaks

character list of created RDS files
}
\description{
Read Peaks for Phenotypes by Class

Create Peaks Directory with Phenotypes by Class
}
