% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno.R
\name{read_pheno}
\alias{read_pheno}
\title{Read Phenotypes}
\usage{
read_pheno(project_df, class, columns = NULL, peak_df = NULL, legacy = FALSE)
}
\arguments{
\item{project_df}{dataframe of project information}

\item{class}{name of class}

\item{columns}{name of columns from phenotype file}

\item{legacy}{use legacy data if `TRUE`}
}
\value{
dataframe
}
\description{
Read Phenotypes
}
