% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project.R
\name{read_project}
\alias{read_project}
\title{Read Project Data}
\usage{
read_project(project_info, dataname, columns, rownames = TRUE, filetype)
}
\arguments{
\item{project_info}{table of project information}

\item{dataname}{name of data object to read}

\item{columns}{columns to select from data object}

\item{rownames}{row names to filter from data object (all if \code{TRUE})}

\item{filetype}{type of file (one of c("fst","rds","csv"))}
}
\value{
data frame with \code{columns} and \code{rownames}.
}
\description{
Read Project Data
}
