% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_query_rds.R
\name{read_query_rds}
\alias{read_query_rds}
\title{Read Query for RDS Object}
\usage{
read_query_rds(project_info, filename)
}
\arguments{
\item{project_info}{table of project information}

\item{filename}{name of RDS file}
}
\value{
data frame read from RDS object.
}
\description{
Read Query for RDS Object
}
