% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_covar.R
\name{scan1_covar}
\alias{scan1_covar}
\title{scan1 for multiple traits with possibly different covariates}
\usage{
scan1_covar(phe_mx, cov_df, probs_obj, kinship, analyses_df, ...)
}
\arguments{
\item{phe_mx}{matrix of phenotypes}

\item{cov_df}{data frame of covariates}

\item{kinship}{kinship matrix or list of kinship matrices}

\item{analyses_df}{data frame of analyses information}
}
\description{
scan1 for multiple traits with possibly different covariates
}
