% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_pattern.R
\name{scan1_pattern}
\alias{scan1_pattern}
\title{scan1 of pattern when multiple traits may have possibly different covariates}
\usage{
scan1_pattern(pheno, phe_df, cov_mx, probs36_obj, K_chr, analyses_df, pats,
  blups)
}
\description{
scan1 of pattern when multiple traits may have possibly different covariates
}
