% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1covar.R
\name{scan1covar}
\alias{scan1covar}
\title{Scan1 for traits with same covariates.}
\usage{
scan1covar(
  pheno_mx,
  covar_df,
  genoprobs,
  kinship,
  peaks_df,
  model = "normal",
  ...
)
}
\arguments{
\item{pheno_mx}{matrix of phenotypes}

\item{covar_df}{data frame of covariates}

\item{genoprobs}{object with genotype probabilities}

\item{kinship}{kinship matrix or list of kinship matrices}

\item{peaks_df}{data frame of peaks information}

\item{model}{model argument for \code{\link[qtl2]{scan1}}}

\item{...}{additional arguments passed on}
}
\value{
object of class \code{\link[qtl2]{scan1}}.
}
\description{
Modified from qtl2mediate version for newer data organization.
Could build out a companion routine splitting on covariates.
}
