% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanInput}
\alias{scanOutput}
\title{Shiny Scan Panel App}
\usage{
scanApp()

scanServer(id, hotspot_list, snp_list, probs_obj, project_df)

scanInput(id)

scanOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Scan Panel App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
