% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanDataApp.R
\name{scanDataApp}
\alias{scanDataApp}
\alias{scanDataServer}
\alias{scanDataInput}
\alias{scanDataOutput}
\title{Shiny Scan Module}
\usage{
scanDataApp()

scanDataServer(id, hotspot_list, probs_obj, project_df)

scanDataInput(id)

scanDataOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 LOD and coefficient plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
