% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterApp.R
\name{scatterApp}
\alias{scatterApp}
\alias{scatterServer}
\alias{scatterInput}
\alias{scatterOutput}
\title{Shiny Scatter App}
\usage{
scatterApp()

scatterServer(
  id,
  hotspot_list,
  pattern_list,
  snp_list,
  pairprobs_obj,
  project_df
)

scatterInput(id)

scatterOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, pattern_list, snp_list, pairprobs_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Scatter App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
