% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlotApp.R
\name{scatterPlotApp}
\alias{scatterPlotApp}
\alias{scatterPlotServer}
\alias{scatterPlotInput}
\alias{scatterPlotOutput}
\title{Shiny Scatter Plot App}
\usage{
scatterPlotApp()

scatterPlotServer(
  id,
  plot_df,
  x_label = shiny::reactive("x"),
  y_label = shiny::reactive("y")
)

scatterPlotInput(id)

scatterPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{plot_df}{reactive data frame containing columns: x, y, and optionally sex, diet, geno}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Scatter Plot App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
