% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParApp.R, R/setupApp.R
\name{setParServer}
\alias{setParServer}
\alias{setupApp}
\alias{setupServer}
\alias{setupInput}
\alias{setupUI}
\alias{setupOutput}
\title{Shiny setup module}
\usage{
setParServer(id, project_df)

setupApp()

setupServer(id, set_par, peak_df, pmap_obj, covar_df, project_df)

setupInput(id)

setupUI(id)

setupOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{peak_df, pmap_obj, covar, projects_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for phenotype selection, with interfaces \code{setupInput} and  \code{setupUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
