% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupServer.R
\name{setupServer}
\alias{setupServer}
\alias{setupInput}
\alias{setupUI}
\title{Shiny setup module}
\usage{
setupServer(
  id,
  pheno_typer,
  peaks_tbl,
  pmap_obj,
  analyses_tbl,
  cov_df,
  projects_info
)

setupInput(id)

setupUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{pheno_typer, peaks_tbl, pmap_obj, analyses_tbl, cov_df, projects_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for phenotype selection, with interfaces \code{setupInput} and  \code{setupUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
