% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAllele.R
\name{shinyAllele}
\alias{shinyAllele}
\title{Shiny coefficient analysis and plot module}
\usage{
shinyAllele(
  id,
  win_par,
  phe_mx,
  cov_df,
  probs_obj,
  K_chr,
  analyses_df,
  patterns,
  scan_pat,
  project_info,
  snp_action
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{win_par, phe_mx, cov_df, probs_obj, K_chr, analyses_df, patterns, scan_pat, project_info, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 coefficient plots, with interfaces \code{shinyAlleleUI} and  \code{shinyAlleleOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
