% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyChrPeak.R
\name{shinyChrPeak}
\alias{shinyChrPeak}
\alias{shinyChrPeakUI}
\title{Shiny chromosome and peak selection}
\usage{
shinyChrPeak(input, output, session, hot_peak = reactive({     NULL }),
  pmap_obj)

shinyChrPeakUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{hot_peak, pmap_obj}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\description{
Shiny module for chromosome and peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

