% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDominance.R
\name{shinyDominance}
\alias{shinyDominance}
\alias{shinyDominanceUI}
\title{Shiny Dominance module}
\usage{
shinyDominance(input, output, session, win_par, phe_df, cov_mx, pheno_anal,
  probs_obj, K_chr)

shinyDominanceUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{plot_type, chr_id, phe_df, cov_mx, pheno_anal, probs_obj, K_chr}{reactive arguments}
}
\description{
Shiny Dominance module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

