% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGeneExon.R
\name{shinyGeneExon}
\alias{shinyGeneExon}
\alias{shinyGeneExonUI}
\title{Shiny Genes and Exons with nearby SNPs module}
\usage{
shinyGeneExon(input, output, session, chr_pos, top_snps_tbl, gene_exon_tbl)

shinyGeneExonUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{chr_pos, top_snps_tbl, gene_exon_tbl}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.

UI for scan1 analyses and summaries to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

