% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHotspot.R
\name{shinyHotspot}
\alias{shinyHotspot}
\alias{shinyHotspotInput}
\alias{shinyHotspotOutput}
\title{Shiny hotspot views}
\usage{
shinyHotspot(input, output, session, set_par, win_par, pheno_type, peaks_tbl,
  pmap_obj, project_info)

shinyHotspotInput(id)

shinyHotspotOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{set_par, win_par, pheno_type, peaks_tbl, pmap_obj, project_info}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\value{
list of inputs and scan summary
}
\description{
Shiny module to view hotspots for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
