% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMediate.R
\name{shinyMediate}
\alias{shinyMediate}
\title{Shiny coefficient analysis and plot module}
\usage{
shinyMediate(
  input,
  output,
  session,
  job_par,
  win_par,
  patterns,
  phe_mx,
  cov_df,
  probs_obj,
  K_chr,
  analyses_df,
  pmap_obj,
  covar,
  analyses_tbl,
  peaks,
  project_info,
  allele_info
)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{job_par, win_par, patterns, phe_mx, cov_df, probs_obj, K_chr, analyses_df, pmap_obj, covar, analyses_tbl, peaks, project_info, allele_info}{reactive arguments}
}
\description{
Shiny module for scan1 coefficient plots, with interfaces \code{shinyMediateUI} and  \code{shinyMediateOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
