% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMediate1.R
\name{shinyMediate1Plot}
\alias{shinyMediate1Plot}
\alias{shinyMediate1PlotUI}
\alias{shinyMediate1PlotOutput}
\title{Shiny coefficient analysis and plot module}
\usage{
shinyMediate1Plot(input, output, session, job_par, win_par, patterns, phe_df,
  cov_mx, probs_obj, K_chr, analyses_df, pmap_obj, datapath)

shinyMediate1PlotUI(id)

shinyMediate1PlotOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{job_par, win_par, patterns, phe_df, cov_mx, probs_obj, K_chr, analyses_df, pmap_obj, datapath}{reactive arguments}

\item{id}{identifier for shiny use}
}
\description{
Shiny module for scan1 coefficient plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
