% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPattern.R
\name{shinyPattern}
\alias{shinyPattern}
\alias{shinyPatternUI}
\title{Shiny Pattern module}
\usage{
shinyPattern(input, output, session, probs1, top_snps_tbl, phe_df, K_chr,
  cov_mx)

shinyPatternUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{plot_type, chr_id, phe_df, cov_mx, pheno_anal, probs_obj, K_chr}{reactive arguments}
}
\description{
Shiny Pattern module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

