% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPhenoPlot.R
\name{shinyPhenoPlot}
\alias{shinyPhenoPlot}
\title{Shiny Phenotype Plot module}
\usage{
shinyPhenoPlot(id, set_par, phe_mx, cov_df)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, phe_mx, cov_df}{reactive arguments}
}
\value{
2-element vector of scan window
}
\description{
Shiny module to plot phenotypes, with interface \code{shinyPhenoPlotUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
