% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProbs.R
\name{shinyProbs}
\alias{shinyProbs}
\alias{shinyProbs36}
\title{Shiny Probability module}
\usage{
shinyProbs(input, output, session, win_par)

shinyProbs36(input, output, session, win_par)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par}{reactive arguments}
}
\description{
Shiny genotype probability access.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

