% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScan1.R
\name{shinyScan1}
\alias{shinyScan1}
\title{Shiny coefficient analysis and plot module}
\usage{
shinyScan1(input, output, session, job_par, win_par, phe_mx, cov_df, probs_obj,
  K_chr, analyses_df, project_info, allele_info)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{job_par, win_par, phe_mx, cov_df, probs_obj, K_chr, analyses_df, project_info, allele_info}{reactive arguments}

\item{id}{shiny identifier}
}
\description{
Shiny module for scan1 LOD and coefficient plots, with interfaces \code{shinyScan1UI} and  \code{shinyScan1Output}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
