% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScan1SNP.R
\name{shinyScan1SNP}
\alias{shinyScan1SNP}
\title{Shiny scan1 SNP analysis and plot module}
\usage{
shinyScan1SNP(input, output, session, snp_par, chr_pos, pheno_names,
  snp_scan_obj, snpinfo, snp_action = shiny::reactive({     "basic" }))
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_par, chr_pos, pheno_names, snp_scan_obj, snpinfo, snp_action}{reactive arguments}

\item{id}{shiny identifier}
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{shinyScan1SNPUI} and  \code{shinyScan1SNPOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
