% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScanCoef.R
\name{shinyScanCoef}
\alias{shinyScanCoef}
\title{Shiny coefficient analysis and plot module}
\usage{
shinyScanCoef(
  id,
  job_par,
  win_par,
  phe_mx,
  cov_df,
  probs_obj,
  K_chr,
  analyses_df,
  project_info,
  allele_info
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{job_par, win_par, phe_mx, cov_df, probs_obj, K_chr, analyses_df, project_info, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 LOD and coefficient plots, with interfaces \code{shinyScanCoefUI} and  \code{shinyScanCoefOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
