% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScatter.R
\name{shinyScatterPlot}
\alias{shinyScatterPlot}
\alias{shinyScatterPlotUI}
\alias{shinyScatterPlotOutput}
\title{Shiny scatter plot module}
\usage{
shinyScatterPlot(input, output, session, med_par, patterns, geno_max, peak_mar,
  med_ls, mediate_obj, phe_df, cov_mx, K_chr)

shinyScatterPlotUI(id)

shinyScatterPlotOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{med_par, patterns, geno_max, peak_mar, med_ls, mediate_obj, phe_df, cov_mx, K_chr}{reactive arguments}

\item{id}{identifier for shiny use}
}
\description{
Shiny module for scatter plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
