% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNP.R
\name{shinySNP}
\alias{shinySNP}
\alias{shinySNPUI}
\title{Shiny SNP analysis and plot module}
\usage{
shinySNP(input, output, session, chr_pos, top_snps_tbl)

shinySNPUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{chr_pos, top_snps_tbl}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.

UI for scan1 analyses and summaries to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

