% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNPAssoc.R
\name{shinySNPAssoc}
\alias{shinySNPAssoc}
\title{Shiny SNP Association}
\usage{
shinySNPAssoc(input, output, session, snp_par, chr_pos, pheno_names,
  snp_scan_obj, snpinfo, top_snps_tbl, gene_exon_tbl, project_info,
  snp_action = shiny::reactive({     "basic" }))
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_par, chr_pos, pheno_names, snp_scan_obj, snpinfo, top_snps_tbl, gene_exon_tbl, project_info, snp_action}{reactive arguments}

\item{id}{shiny identifier}
}
\value{
tbl with top SNPs
}
\description{
Shiny module for SNP association mapping, with interfaces \code{shinySNPAssocInput}, \code{shinySNPAssocUI} and  \code{shinySNPAssocOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
