% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNPCsq.R
\name{shinySNPCsq}
\alias{shinySNPCsq}
\alias{shinySNPCsqOutput}
\alias{shinySNPCsqUI}
\title{Shiny SNP Consequence}
\usage{
shinySNPCsq(input, output, session, snp_scan_obj, chr_pos,
  snp_action = reactive({     "basic" }))

shinySNPCsqUI(id)

shinySNPCsqOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_scan_obj, chr_pos, snp_action}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\value{
tbl with top SNPs
}
\description{
Shiny SNP Consequence
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

