% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNPGene.R
\name{shinySNPGene}
\alias{shinySNPGene}
\title{Shiny SNP Association}
\usage{
shinySNPGene(
  id,
  snp_par,
  chr_pos,
  pheno_names,
  snp_scan_obj,
  snpinfo,
  top_snps_tbl,
  gene_exon_tbl,
  project_info,
  snp_action = shiny::reactive({
     "basic"
 })
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_par, chr_pos, pheno_names, snp_scan_obj, snpinfo, top_snps_tbl, gene_exon_tbl, project_info, snp_action}{reactive arguments}
}
\value{
tbl with top SNPs
}
\description{
Shiny module for SNP association mapping, with interfaces \code{shinySNPGeneInput}, \code{shinySNPGeneUI} and  \code{shinySNPGeneOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
