% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNPSetup.R
\name{shinySNPSetup}
\alias{shinySNPSetup}
\title{Shiny SNP and Allele analysis and plot module}
\usage{
shinySNPSetup(
  input,
  output,
  session,
  job_par,
  win_par,
  phe_mx,
  cov_df,
  K_chr,
  analyses_df,
  project_info,
  allele_info,
  snp_action = shiny::reactive({     "basic" })
)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{job_par, win_par, phe_mx, cov_df, K_chr, analyses_df, project_info, allele_info, snp_action}{reactive arguments}
}
\description{
Shiny module to coordinate SNP and allele analyses and plots, with interfaces \code{shinySNPSetupUI} and  \code{shinySNPSetupOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
