% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNPSum.R
\name{shinySNPSum}
\alias{shinySNPSum}
\title{Shiny SNP summary module}
\usage{
shinySNPSum(input, output, session, chr_pos, top_snps_tbl, project_info,
  snp_action = shiny::reactive({     "basic" }))
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{chr_pos, top_snps_tbl, project_info, snp_action}{reactive arguments}

\item{id}{shiny identifier}
}
\description{
Shiny module for SNP summary, with interfaces \code{shinySNPInput}, \code{shinySNPUI} and  \code{shinySNPOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
