% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTopFeature.R
\name{shinyTopFeature}
\alias{shinyTopFeature}
\alias{shinyTopFeatureInput}
\alias{shinyTopFeatureOutput}
\alias{shinyTopFeatureUI}
\title{Shiny Top Features in SNP Region module}
\usage{
shinyTopFeature(input, output, session, snp_par, chr_pos, snp_scan_obj,
  top_snps_tbl, gene_exon_tbl, snp_action = reactive({     "basic" }))

shinyTopFeatureInput(id)

shinyTopFeatureUI(id)

shinyTopFeatureOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_scan_obj, top_snps_tbl, gene_exon_tbl, snp_action}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

