% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyWindow.R
\name{shinyWindow}
\alias{shinyWindow}
\alias{shinyWindowUI}
\title{Shiny window selection}
\usage{
shinyWindow(input, output, session, pmap_obj, hot_peak = reactive({     NULL
  }))

shinyWindowUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{pmap_obj}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\description{
Shiny module for window selection around key peak.

UI for window selection to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

