% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpFeatureApp.R
\name{snpFeatureApp}
\alias{snpFeatureApp}
\alias{snpFeatureServer}
\alias{snpFeatureOutput}
\title{Shiny SNP Features in SNP Region module}
\usage{
snpFeatureApp()

snpFeatureServer(id, snp_list)

snpFeatureOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list}{reactive arguments}
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{snpFeatureInput}, \code{snpFeatureUI} and  \code{snpFeatureOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
