% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpGeneApp.R
\name{snpGeneApp}
\alias{snpGeneApp}
\alias{snpGeneServer}
\alias{snpGeneInput}
\alias{snpGeneOutput}
\title{Shiny SNP Gene App}
\usage{
snpGeneApp()

snpGeneServer(id, hotspot_list, snp_list, project_df)

snpGeneInput(id)

snpGeneOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list, project_df}{reactive arguments}
}
\value{
tbl with top SNPs
}
\description{
Shiny module for SNP association mapping, with interfaces \code{snpGeneInput}, \code{snpGeneUI} and  \code{snpGeneOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
