% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpListApp.R
\name{snpListApp}
\alias{snpListApp}
\alias{snpListServer}
\alias{snpListInput}
\alias{snpListInput2}
\alias{snpListUI}
\alias{snpListOutput}
\title{Shiny SNP List App}
\usage{
snpListApp()

snpListServer(
  id,
  hap_par,
  win_par,
  peak_df,
  pheno_mx,
  covar_df,
  K_chr,
  project_df,
  snp_action = shiny::reactive({
     "basic"
 })
)

snpListInput(id)

snpListInput2(id)

snpListUI(id)

snpListOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hap_par, win_par, peak_df, pheno_mx, covar_df, K_chr, project_df, snp_action}{reactive arguments}
}
\description{
Shiny module to create list of SNP objects.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
