% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpPatternApp.R
\name{snpPatternApp}
\alias{snpPatternApp}
\alias{snpPatternServer}
\alias{snpPatternInput}
\alias{snpPatternUI}
\alias{snpPatternOutput}
\title{Shiny SNP Pattern App}
\usage{
snpPatternApp()

snpPatternServer(id, snp_list, allele_info)

snpPatternInput(id)

snpPatternUI(id)

snpPatternOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list, allele_info}{reactive arguments}
}
\description{
Shiny SNP Pattern App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
