% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpPatternServer.R
\name{snpPatternServer}
\alias{snpPatternServer}
\alias{snpPatternInput}
\alias{snpPatternUI}
\alias{snpPatternOutput}
\title{Shiny top SNP analysis and plot module}
\usage{
snpPatternServer(
  id,
  snp_par,
  chr_pos,
  pheno_names,
  snp_scan_obj,
  snpinfo,
  top_snps_tbl,
  gene_exon_tbl,
  allele_info,
  snp_action = shiny::reactive({
     "basic"
 })
)

snpPatternInput(id)

snpPatternUI(id)

snpPatternOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_par, chr_pos, pheno_names, snp_scan_obj, snpinfo, top_snps_tbl, gene_exon_tbl, allele_info, snp_action}{reactive arguments}
}
\description{
Shiny module for top SNP analysis and plots, with interfaces \code{snpPatternInput}, \code{snpPatternUI} and  \code{snpPatternOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
