% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpPlotServer.R
\name{snpPlotServer}
\alias{snpPlotServer}
\alias{snpPlotUI}
\alias{snpPlotOutput}
\title{Shiny SNP plot module}
\usage{
snpPlotServer(id, snp_list)

snpPlotUI(id)

snpPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list}{reactive arguments}
}
\description{
Shiny SNP plot module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
