% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpSetupApp.R
\name{snpSetupApp}
\alias{snpSetupApp}
\alias{snpSetupServer}
\alias{snpSetupInput}
\alias{snpSetupUI}
\alias{snpSetupOutput}
\title{Shiny SNP and Allele analysis and plot App}
\usage{
snpSetupApp()

snpSetupServer(
  id,
  hap_par,
  win_par,
  peak_df,
  pheno_mx,
  covar_df,
  K_chr,
  allele_info,
  project_df,
  snp_action = shiny::reactive({
     "basic"
 })
)

snpSetupInput(id)

snpSetupUI(id)

snpSetupOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hap_par, win_par, peak_df, pheno_mx, covar_df, K_chr, allele_info, project_df, snp_action}{reactive arguments}
}
\description{
Shiny module to coordinate SNP and allele analyses and plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
