% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpSetupServer.R
\name{snpSetupServer}
\alias{snpSetupServer}
\alias{snpSetupUI}
\alias{snpSetupOutput}
\title{Shiny SNP and Allele analysis and plot module}
\usage{
snpSetupServer(
  id,
  job_par,
  win_par,
  phe_mx,
  cov_df,
  K_chr,
  analyses_df,
  project_df,
  allele_info,
  snp_action = shiny::reactive({
     "basic"
 })
)

snpSetupUI(id)

snpSetupOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{job_par, win_par, phe_mx, cov_df, K_chr, analyses_df, project_df, allele_info, snp_action}{reactive arguments}
}
\description{
Shiny module to coordinate SNP and allele analyses and plots, with interfaces \code{snpSetupUI} and  \code{snpSetupOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
