% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpTableApp.R
\name{snpTableApp}
\alias{snpTableApp}
\alias{snpTableServer}
\alias{snpTableOutput}
\title{Shiny SNP summary module}
\usage{
snpTableApp()

snpTableServer(id, snp_list, project_df)

snpTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for SNP summary, with interfaces \code{shinySNPInput}, \code{shinySNPUI} and  \code{shinySNPOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
