% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_na.R
\name{summary_na}
\alias{summary_na}
\title{Summary across traits}
\usage{
summary_na(phe)
}
\arguments{
\item{phe}{data frame of phenotypes}
}
\description{
Summary taking care to include NA for all traits
}
\examples{
\dontrun{summary_na(phe)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

