% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_snp_asso.R
\name{top_snp_asso}
\alias{top_snp_asso}
\title{Top SNP Association Plot Wrapper}
\usage{
top_snp_asso(pheno, scan_obj, xlim, snp_action = "basic", drop.hilit = NULL,
  main = mytitle, pch = 1, cex = 0.75, col = "#7570b3",
  col.hilit = "#d95f02", ...)
}
\arguments{
\item{pheno}{name of phenotype for effect scan}

\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{xlim}{x limits for plot}

\item{snp_action}{character string for plot}

\item{drop.hilit}{hilight SNPs within this value of max}

\item{main}{plot title}

\item{pch, cex, col, col.hilit}{plot parameters}

\item{...}{additional plot parameters}
}
\description{
Top SNP Association Plot Wrapper
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}

