% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triadApp.R
\name{triadApp}
\alias{triadApp}
\alias{triadServer}
\alias{triadInput}
\alias{triadOutput}
\title{Shiny Triad App}
\usage{
triadApp()

triadServer(id, hotspot_list, snp_list, mediate_list, probs_obj)

triadInput(id)

triadOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, snp_list, mediate_list, probs_obj}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Triad App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
