% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triadServer.R
\name{triadServer}
\alias{triadServer}
\alias{triadUI}
\alias{triadOutput}
\title{Shiny triad plot module}
\usage{
triadServer(
  id,
  med_par,
  patterns,
  geno_max,
  peak_mar,
  med_ls,
  mediate_obj,
  phe_mx,
  cov_df,
  K_chr,
  probs_obj,
  chr_id,
  sdp
)

triadUI(id)

triadOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{med_par, patterns, geno_max, peak_mar, med_ls, mediate_obj, phe_mx, cov_df, K_chr, probs_obj, chr_id, sdp}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for triad plots, with interfaces \code{triadUI} and  \code{triadOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
