% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winParApp.R
\name{winParApp}
\alias{winParApp}
\alias{winParServer}
\alias{winParInput}
\alias{winParOutput}
\title{Shiny Window Parameters App}
\usage{
winParApp()

winParServer(id, hotspot_df, project_df)

winParInput(id)

winParOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_df, project_df}{reactive arguments}
}
\value{
list of inputs and scan summary
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
