% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashServer.R
\name{dashServer}
\alias{dashServer}
\alias{dashInput}
\alias{dashUI}
\alias{dashOutput}
\alias{dashOutput2}
\title{Shiny dash module for qtl2}
\usage{
dashServer(id, projects_df)

dashInput(id)

dashUI(id)

dashOutput(id)

dashOutput2(id)
}
\arguments{
\item{id}{shiny identifier}

\item{projects_df}{static data frame with project information}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for phenotype selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
