% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExonServer.R
\name{geneExonServer}
\alias{geneExonServer}
\alias{geneExonInput}
\alias{geneExonUI}
\alias{geneExonOutput}
\title{Shiny Genes and Exons with nearby SNPs module}
\usage{
geneExonServer(id, snp_list)

geneExonInput(id)

geneExonUI(id)

geneExonOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{geneExonInput}, \code{geneExonUI} and  \code{geneExonOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
