% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneRegionServer.R
\name{geneRegionServer}
\alias{geneRegionServer}
\alias{geneRegionInput}
\alias{geneRegionUI}
\alias{geneRegionOutput}
\title{Shiny Genes in SNP Region module}
\usage{
geneRegionServer(
  id,
  snp_par,
  top_snps_tbl,
  project_df,
  snp_action = shiny::reactive({
     "basic"
 })
)

geneRegionInput(id)

geneRegionUI(id)

geneRegionOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_par, top_snps_tbl, project_df, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{geneRegionInput}, \code{geneRegionUI} and  \code{geneRegionOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
