% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploApp.R
\name{haploApp}
\alias{haploApp}
\alias{haploServer}
\alias{haploUI}
\alias{haploOutput}
\title{Shiny Haplotype Analysis App}
\usage{
haploApp()

haploServer(id, hotspot_list, project_df)

haploUI(id)

haploOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for analysis based on haplotype alleles, with interface \code{haploUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
