% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotServer.R
\name{hotspotApp}
\alias{hotspotApp}
\alias{hotspotServer}
\alias{hotspotInput}
\alias{hotspotOutput}
\title{Shiny hotspot module}
\usage{
hotspotApp()

hotspotServer(id, set_par, pheno_type, peaks_tbl, pmap_obj, project_info)

hotspotInput(id)

hotspotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, pheno_type, peaks_tbl, pmap_obj, project_info}{reactive arguments}
}
\value{
list of inputs and scan summary

No return value; called for side effects.
}
\description{
Shiny module to view hotspots for peak selection, with interfaces \code{hotspotInput} and  \code{hotspotOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
