% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotApp.R
\name{hotspotApp}
\alias{hotspotApp}
\alias{hotspotServer}
\alias{hotspotInput}
\alias{hotspotUI}
\alias{hotspotOutput}
\title{Shiny Hotspot App}
\usage{
hotspotApp()

hotspotServer(id, set_par, peak_df, pmap_obj, project_df)

hotspotInput(id)

hotspotUI(id)

hotspotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\value{
list of inputs and scan summary

No return value; called for side effects.
}
\description{
Shiny module to view hotspots for peak selection, with interfaces \code{hotspotInput} and  \code{hotspotOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
