% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_chr_pos}
\alias{make_chr_pos}
\title{Make chr_pos from chr, peak and window}
\usage{
make_chr_pos(chr_id = NULL, peak_Mbp = NULL, window_Mbp = NULL,
  range = c(peak_Mbp - window_Mbp, peak_Mbp + window_Mbp))
}
\arguments{
\item{chr_id}{chromosome ID}

\item{peak_Mbp}{position of peak in Mbp}

\item{window_Mbp}{half-width of window around peak in Mbp}

\item{range}{left and right window positions in Mbp}
}
\description{
Create character string of chr:left-right with left=peak-window, right=peak+window
}
\examples{
\dontrun{make_chr_pos(chr_id, peak_Mbp, window_Mbp)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

