% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediateApp.R
\name{mediateApp}
\alias{mediateApp}
\alias{mediateServer}
\alias{mediateInput}
\alias{mediateUI}
\alias{mediateOutput}
\title{Shiny Mediate App}
\usage{
mediateApp()

mediateServer(id, hotspot_list, snp_list, probs_obj, project_df)

mediateInput(id)

mediateUI(id)

mediateOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, patterns, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny Mediate App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
