% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediateServer.R
\name{mediateApp}
\alias{mediateApp}
\alias{mediateServer}
\alias{mediateUI}
\alias{mediateOutput}
\title{Shiny coefficient analysis and plot module}
\usage{
mediateApp()

mediateServer(id, hotspot_list, probs_obj, patterns, project_df)

mediateUI(id)

mediateOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, patterns, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 coefficient plots, with interfaces \code{mediateUI} and  \code{mediateOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
