% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternApp.R, R/patternPlotApp.R
\name{patternApp}
\alias{patternApp}
\alias{patternServer}
\alias{patternInput}
\alias{patternUI}
\alias{patternOutput}
\alias{patternPlotServer}
\alias{patternPlotOutput}
\title{Shiny Pattern App}
\usage{
patternApp()

patternServer(
  id,
  hotspot_list,
  dip_par,
  pairprobs_obj,
  patterns,
  snp_action = shiny::reactive({
     "basic"
 }),
  project_df
)

patternInput(id)

patternUI(id)

patternOutput(id)

patternPlotServer(id, pattern_list, pairprobs_obj)

patternPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, dip_par, pairprobs_obj, patterns, snp_action, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for SNP pattern plots, with interfaces \code{patternUI} and  \code{patternOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
