% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakApp.R
\name{peakApp}
\alias{peakApp}
\alias{peakServer}
\alias{peakInput}
\alias{peakUI}
\alias{peakOutput}
\title{Shiny peaks module}
\usage{
peakApp()

peakServer(id, set_par, pheno_type, peaks_tbl, pmap_obj, project_info)

peakInput(id)

peakUI(id)

peakOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, pheno_type, peaks_tbl, pmap_obj, project_info}{reactive arguments}
}
\value{
list of inputs and scan summary
}
\description{
Shiny module for peaks selection, with interfaces \code{peakInput}, \code{peakUI} and  \code{peakOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
