% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoApp.R
\name{phenoApp}
\alias{phenoApp}
\alias{phenoServer}
\alias{phenoUI}
\alias{phenoOutput}
\title{Shiny Phenotype App}
\usage{
phenoApp()

phenoServer(id, set_par, pheno_names, project_df)

phenoUI(id)

phenoOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
