% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoApp.R
\name{phenoApp}
\alias{phenoApp}
\alias{phenoServer}
\alias{phenoInput}
\alias{phenoOutput}
\title{Shiny phenotype selection}
\usage{
phenoApp()

phenoServer(id, set_par, win_par, peak_df, covar_df, project_df)

phenoInput(id)

phenoOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, win_par, peak_df, analyses_df, covar, project_df}{reactive arguments}
}
\description{
Shiny module for phenotype selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
