% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoReadApp.R
\name{phenoReadApp}
\alias{phenoReadApp}
\alias{phenoReadServer}
\alias{phenoReadOutput}
\title{Shiny Pheno Read App}
\usage{
phenoReadApp()

phenoReadServer(id, set_par, pheno_names, project_df)

phenoReadOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, peak_df, pmap_obj, project_df}{reactive arguments}
}
\description{
Shiny module for peak selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
