% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eff.R
\name{plot_eff}
\alias{plot_eff}
\title{Plot effects}
\usage{
plot_eff(pheno, eff_obj, scan_obj, xlim = NULL, chr_id = 1,
  addlod = FALSE)
}
\arguments{
\item{pheno}{name of phenotype for effect scan}

\item{eff_obj}{object of class \code{\link[qtl2pattern]{listof_scan1coef}}}

\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{xlim}{x limits for plot}

\item{chr_id}{chromosome ID}
}
\description{
Plot of effects.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
