% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_eff}
\alias{plot_eff}
\title{Plot effects}
\usage{
plot_eff(pheno, scan_obj, eff_obj, xlim)
}
\arguments{
\item{pheno}{name of phenotype for effect scan}

\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{eff_obj}{object of class \code{\link[doqtl2]{listof_scan1coefCC}}}

\item{xlim}{x limits for plot}
}
\description{
Plot of effects.
}
\examples{
\dontrun{plot_eff(pheno, scan_obj, eff_obj, xlim)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}

