% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project.R
\name{project_classes}
\alias{project_classes}
\title{Project Classes}
\usage{
project_classes(project_df)
}
\arguments{
\item{project_df}{}
}
\value{
character list of classes
}
\description{
Find names of project classes using phenotype files.
}
