% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2shinyApp.R
\name{qtl2shinyApp}
\alias{qtl2shinyApp}
\alias{qtl2shinyServer}
\alias{qtl2shinyUI}
\title{Shiny QTL2 Shiny App}
\usage{
qtl2shinyApp()

qtl2shinyServer(id, projects_df)

qtl2shinyUI(id)
}
\arguments{
\item{id}{shiny identifier}

\item{projects_df}{static data frame with project information}
}
\value{
No return value; called for side effects.
}
\description{
Shiny QTL2 Shiny App
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
