% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2shinyApp.R
\name{qtl2shinyApp}
\alias{qtl2shinyApp}
\title{Shiny app for doqtl2}
\usage{
qtl2shinyApp(appfile = "app.R")
}
\arguments{
\item{appfile}{name of app file for shiny}
}
\value{
No return value; called for side effects.
}
\description{
run shiny app for doqtl2
}
\examples{
\dontrun{doqtl2_app()}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
