% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project.R
\name{read_project}
\alias{read_project}
\title{Read Project Data}
\usage{
read_project(
  project_df,
  dataname,
  columns,
  rownames = TRUE,
  filetype,
  class = NULL,
  ...
)
}
\arguments{
\item{project_df}{table of project information}

\item{dataname}{name of data object to read}

\item{columns}{columns to select from data object}

\item{rownames}{row names to filter from data object (all if \code{TRUE})}

\item{filetype}{type of file (one of c("fst","rds","csv"))}

\item{class}{name of subdirectory if not `NULL`}

\item{...}{additional parameters}
}
\value{
data frame with \code{columns} and \code{rownames}.
}
\description{
Read Project Data
}
