% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project.R
\name{read_project}
\alias{read_project}
\title{Read Project Data}
\usage{
read_project(
  project_df,
  dataname,
  folder = NULL,
  columns = NULL,
  rownames = TRUE,
  ...,
  filetype,
  legacy = FALSE
)
}
\arguments{
\item{project_df}{table of project information}

\item{dataname}{name of data object to read}

\item{folder}{folder containing data if not `NULL`}

\item{columns}{columns to select from data object}

\item{rownames}{row names to filter from data object (all if \code{TRUE})}

\item{...}{additional parameters}

\item{filetype}{type of file (one of c("fst","rds","csv"))}

\item{legacy}{look in `legacy` folder for old data if `TRUE`}
}
\value{
data frame with \code{columns} and \code{rownames}.
}
\description{
Read Project Data
}
