% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_covar.R
\name{scan1_covar}
\alias{scan1_covar}
\title{scan1 for multiple traits with possibly different covariates}
\usage{
scan1_covar(probs_obj, phe_df, K_chr, cov_mx, analyses_df)
}
\description{
scan1 for multiple traits with possibly different covariates
}
