% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_covar.R
\name{scan1_covar}
\alias{scan1_covar}
\alias{which_covar}
\alias{covar_df_mx}
\alias{wh_covar}
\alias{sexcovar}
\title{Scan1 for multiple traits with possibly different covariates, and some helper routines.}
\usage{
scan1_covar(phe_mx, cov_df, probs_obj, kinship, analyses_df, ...)

which_covar(analyses_df)

covar_df_mx(addcovar)

wh_covar(analyses_df, wh, cov_df)

sexcovar(addcovar, sex_type = "all")
}
\arguments{
\item{phe_mx}{matrix of phenotypes}

\item{cov_df}{data frame of covariates}

\item{probs_obj}{object with genotype probabilities}

\item{kinship}{kinship matrix or list of kinship matrices}

\item{analyses_df}{data frame of analyses information}

\item{...}{additional arguments passed on}

\item{addcovar}{Data frame of additive covariate.}

\item{wh}{Index for which row of \code{analyses_df} to use.}

\item{sex_type}{subset by sex if \code{"F"} or \code{"M"}.}
}
\value{
object of class \code{\link[qtl2]{scan1}}.
}
\description{
Scan1 for multiple traits with possibly different covariates, and some helper routines.
}
