% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanUI}
\alias{scanOutput}
\title{Shiny Scan Module}
\usage{
scanApp()

scanServer(
  id,
  job_par,
  win_par,
  phe_mx,
  cov_df,
  probs_obj,
  K_chr,
  analyses_df,
  project_df,
  allele_info
)

scanUI(id)

scanOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{job_par, win_par, phe_mx, cov_df, probs_obj, K_chr, analyses_df, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 LOD and coefficient plots, with interfaces \code{scanUI} and  \code{scanOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
