% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanInput}
\alias{scanUI}
\alias{scanOutput}
\title{Shiny Scan Module}
\usage{
scanApp()

scanServer(id, hotspot_list, probs_obj, project_df)

scanInput(id)

scanUI(id)

scanOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hotspot_list, probs_obj, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 LOD and coefficient plots, with interfaces \code{scanUI} and  \code{scanOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
