% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanUI}
\alias{scanOutput}
\title{Shiny Scan Module}
\usage{
scanApp()

scanServer(
  id,
  hap_par,
  win_par,
  peak_df,
  pheno_mx,
  covar_df,
  K_chr,
  probs_obj,
  allele_info,
  project_df
)

scanUI(id)

scanOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{hap_par, hotspot_list, pheno_mx, covar_df, probs_obj, K_chr, project_df, allele_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 LOD and coefficient plots, with interfaces \code{scanUI} and  \code{scanOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
