% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParApp.R
\name{setParApp}
\alias{setParApp}
\alias{setParInput}
\alias{setParOutput}
\title{Shiny Setup Parameters App}
\usage{
setParApp()

setParInput(id)

setParOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{peak_df, pmap_obj, analyses_df, covar, projects_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for phenotype selection, with interfaces \code{setupInput} and  \code{setupUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
