% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupApp.R
\name{setupApp}
\alias{setupApp}
\alias{setupServer}
\alias{setupInput}
\alias{setupUI}
\alias{setupOutput}
\title{Shiny setup module}
\usage{
setupApp()

setupServer(id, peak_df, pmap_obj, analyses_tbl, cov_df, project_info)

setupInput(id)

setupUI(id)

setupOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{peak_df, pmap_obj, analyses_tbl, cov_df, projects_info}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for phenotype selection, with interfaces \code{setupInput} and  \code{setupUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
