% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAllelePat.R
\name{shinyAllelePat}
\alias{shinyAllelePat}
\alias{shinyAllelePatInput}
\alias{shinyAllelePatUI}
\alias{shinyAllelePatOutput}
\title{Shiny top SNP analysis and plot module}
\usage{
shinyAllelePat(input, output, session, snp_par, chr_pos, pheno_names,
  snp_scan_obj, snpinfo, top_snps_tbl, gene_exon_tbl,
  snp_action = shiny::reactive({     "basic" }))

shinyAllelePatInput(id)

shinyAllelePatUI(id)

shinyAllelePatOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_par, chr_pos, pheno_names, snp_scan_obj, snpinfo, top_snps_tbl, gene_exon_tbl, snp_action}{reactive arguments}

\item{id}{identifier for \code{\link{shinyAllelePat}} use}
}
\description{
Shiny module for top SNP analysis and plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
