% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDiplo.R
\name{shinyDiplo}
\alias{shinyDiplo}
\alias{shinyDiploUI}
\title{Shiny Diplotype module}
\usage{
shinyDiplo(input, output, session, win_par, phe_df, cov_mx, K_chr, analyses_df,
  data_path)

shinyDiploUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par, phe_df, cov_mx, K_chr, analyses_df, data_path}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny diplotype SNP/Gene action analysis.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
