% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDiplo.R
\name{shinyDiplo}
\alias{shinyDiplo}
\title{Shiny Diplotype module}
\usage{
shinyDiplo(
  input,
  output,
  session,
  win_par,
  phe_mx,
  cov_df,
  K_chr,
  analyses_df,
  project_info,
  allele_info
)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par, phe_mx, cov_df, K_chr, analyses_df, project_info, allele_info}{reactive arguments}
}
\description{
Shiny diplotype SNP/Gene action analysis, with interface \code{shinyDiploUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
