% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGeneExon.R
\name{shinyGeneExon}
\alias{shinyGeneExon}
\alias{shinyGeneExonInput}
\alias{shinyGeneExonUI}
\alias{shinyGeneExonOutput}
\title{Shiny Genes and Exons with nearby SNPs module}
\usage{
shinyGeneExon(input, output, session, snp_par, chr_pos, top_snps_tbl,
  gene_exon_tbl, snp_action = shiny::reactive({     "basic" }))

shinyGeneExonInput(id)

shinyGeneExonUI(id)

shinyGeneExonOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_par, chr_pos, top_snps_tbl, gene_exon_tbl, snp_action}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
