% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGeneExon.R
\name{shinyGeneExon}
\alias{shinyGeneExon}
\title{Shiny Genes and Exons with nearby SNPs module}
\usage{
shinyGeneExon(
  id,
  snp_par,
  chr_pos,
  top_snps_tbl,
  gene_exon_tbl,
  snp_action = shiny::reactive({
     "basic"
 })
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_par, chr_pos, top_snps_tbl, gene_exon_tbl, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{shinyGeneExonInput}, \code{shinyGeneExonUI} and  \code{shinyGeneExonOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
