% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGeneRegion.R
\name{shinyGeneRegion}
\alias{shinyGeneRegion}
\alias{shinyGeneRegionUI}
\title{Shiny Genes in SNP Region module}
\usage{
shinyGeneRegion(input, output, session, top_snps_tbl, feature_file)

shinyGeneRegionUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{top_snps_tbl, feature_file}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.

UI for scan1 analyses and summaries to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

