% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGeneRegion.R
\name{shinyGeneRegion}
\alias{shinyGeneRegion}
\alias{shinyGeneRegionInput}
\alias{shinyGeneRegionUI}
\alias{shinyGeneRegionOutput}
\title{Shiny Genes in SNP Region module}
\usage{
shinyGeneRegion(input, output, session, snp_par, top_snps_tbl, project_info,
  snp_action = shiny::reactive({     "basic" }))

shinyGeneRegionInput(id)

shinyGeneRegionUI(id)

shinyGeneRegionOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_par, top_snps_tbl, project_info}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
