% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHaplo.R
\name{shinyHaplo}
\alias{shinyHaplo}
\alias{shinyHaploUI}
\title{Shiny haplotype analysis}
\usage{
shinyHaplo(input, output, session, win_par, pmap_obj, phe_df, cov_mx, K_chr,
  analyses_df)

shinyHaploUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par, pmap_obj, phe_df, cov_mx, K_chr, analyses_df}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for analysis based on haplotype alleles.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
