% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMain.R
\name{shinyMain}
\alias{shinyMain}
\alias{shinyMainInput}
\alias{shinyMainUI}
\alias{shinyMainOutput}
\alias{shinyMainOutput2}
\title{Shiny app for DOQTL selection}
\usage{
shinyMain(input, output, session, projects_info)

shinyMainInput(id)

shinyMainUI(id)

shinyMainOutput(id)

shinyMainOutput2(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{projects_info}{reactive arguments}

\item{id}{shiny identifier}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for phenotype selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
