% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPattern.R
\name{shinyPattern}
\alias{shinyPattern}
\alias{shinyPatternUI}
\title{Shiny Pattern module}
\usage{
shinyPattern(input, output, session, probs1, patterns, phe_df, K_chr, cov_mx,
  chr_pos, snp_action = reactive({     NULL }))

shinyPatternUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{probs1, patterns, phe_df, K_chr, cov_mx, chr_pos, snp_action}{reactive arguments}
}
\description{
Shiny Pattern module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

