% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPattern.R
\name{shinyPattern}
\alias{shinyPattern}
\title{Shiny Pattern module}
\usage{
shinyPattern(
  id,
  job_par,
  chr_pos,
  win_par,
  phe_mx,
  cov_df,
  pairprobs_obj,
  K_chr,
  analyses_df,
  patterns,
  project_info,
  allele_info,
  snp_action = shiny::reactive({
     NULL
 })
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{job_par, chr_pos, win_par, phe_mx, cov_df, pairprobs_obj, K_chr, analyses_df, patterns, project_info, allele_info, snp_action}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for SNP pattern plots, with interfaces \code{shinyPatternUI} and  \code{shinyPatternOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
