% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPattern.R
\name{shinyPattern}
\alias{shinyPattern}
\alias{shinyPatternUI}
\alias{shinyPatternOutput}
\title{Shiny Pattern module}
\usage{
shinyPattern(input, output, session, chr_pos, phe_df, K_chr, cov_mx,
  probs36_obj, patterns, snp_action = shiny::reactive({     NULL }))

shinyPatternUI(id)

shinyPatternOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{chr_pos, phe_df, K_chr, cov_mx, probs36_obj, patterns, snp_action}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny Pattern module
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
