% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPeaks.R
\name{shinyPeaks}
\alias{shinyPeaks}
\alias{shinyPeaksInput}
\alias{shinyPeaksOutput}
\title{Shiny peaks selection}
\usage{
shinyPeaks(input, output, session, pheno_type, peaks_tbl, pmap_obj)

shinyPeaksInput(id)

shinyPeaksOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{pheno_type, peaks_tbl, pmap_obj}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\value{
list of inputs and scan summary
}
\description{
Shiny module for peaks selection. Does not yet have return.

UI for peak selection to use in shiny module.

UI for peak selection to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

