% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPhenoPlot.R
\name{shinyPhenoPlot}
\alias{shinyPhenoPlot}
\alias{shinyPhenoPlotUI}
\title{Shiny Phenotype Plot module}
\usage{
shinyPhenoPlot(input, output, session, raw_phe_df, phe_df, cov_mx)

shinyPhenoPlotUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{raw_phe_df, phe_df, cov_mx}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\value{
2-element vector of scan window
}
\description{
Shiny module to plot phenotypes.

UI for phenotype plots and summaries to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

