% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPhenos.R
\name{shinyPhenos}
\alias{shinyPhenos}
\title{Shiny phenotype selection}
\usage{
shinyPhenos(id, set_par, win_par, peaks_df, analyses_tbl, cov_df, project_info)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{set_par, win_par, peaks_df, analyses_tbl, cov_df, project_info}{reactive arguments}
}
\description{
Shiny module for phenotype selection, with interfaces \code{shinyPhenosUI} and  \code{shinyPhenosOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
