% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPhenos.R
\name{shinyPhenos}
\alias{shinyPhenos}
\alias{shinyPhenosUI}
\alias{shinyPhenosOutput}
\title{Shiny phenotype selection}
\usage{
shinyPhenos(input, output, session, setup_par, peaks_tbl, analyses_tbl,
  chr_peak)

shinyPhenosUI(id)

shinyPhenosOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{setup_par, peaks_tbl, analyses_tbl, chr_peak}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\description{
Shiny module for phenotype selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
