% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPhenos.R
\name{shinyPhenos}
\alias{shinyPhenos}
\alias{shinyPhenosUI}
\title{Shiny phenotype selection}
\usage{
shinyPhenos(input, output, session, pheno_type, peaks_tbl, analyses_tbl,
  hot_peak, win_par)

shinyPhenosUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{pheno_type, peaks_tbl, analyses_tbl, hot_peak, win_par}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\description{
Shiny module for phenotype selection.

UI for phenotype selection to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

