% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProbs.R
\name{shinyProbs}
\alias{shinyProbs}
\alias{shinyPairProbs}
\alias{shinySNPProbs}
\title{Shiny Probability module}
\usage{
shinyProbs(input, output, session, win_par, project_info)

shinyPairProbs(input, output, session, win_par, project_info)

shinySNPProbs(
  input,
  output,
  session,
  win_par,
  pheno_names,
  probs_obj,
  project_info
)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par, pheno_names, probs_obj, project_info}{reactive arguments}
}
\value{
Object of class \code{probs}.
}
\description{
Shiny genotype probability access.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
