% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProbs.R
\name{shinyProbs}
\alias{shinyProbs}
\alias{shinyProbs36}
\alias{shinySNPProbs}
\title{Shiny Probability module}
\usage{
shinyProbs(input, output, session, win_par, data_path)

shinyProbs36(input, output, session, win_par, data_path)

shinySNPProbs(input, output, session, win_par, pheno_names, probs_obj,
  data_path)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par, pheno_names, probs_obj, data_path}{reactive arguments}
}
\description{
Shiny genotype probability access.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

