% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProject.R
\name{shinyProject}
\alias{shinyProject}
\title{Shiny project selection}
\usage{
shinyProject(id, projects_info)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{projects_info}{reactive project info data frame}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for selection of project, with interface \code{shinyProjectUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
