% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProject.R
\name{shinyProject}
\alias{shinyProject}
\title{Shiny project selection}
\usage{
shinyProject(input, output, session, projects_info)
}
\arguments{
\item{input, output, session, projects_info}{standard shiny arguments}

\item{id}{shiny identifier}
}
\description{
Shiny module for selection of project, with interface \code{shinyProjectUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
