% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScan1.R
\name{shinyScan1}
\alias{shinyScan1}
\alias{shinyScan1UI}
\title{Shiny scan1 analysis and plot module}
\usage{
shinyScan1(input, output, session, plot_type, chr_id, phe_df, cov_mx,
  pheno_anal, probs_obj, K_chr)

shinyScan1UI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{plot_type, chr_id, phe_df, cov_mx, pheno_anal, probs_obj, K_chr}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}} use}
}
\value{
object of class \code{\link[qtl2scan]{scan1}}
}
\description{
Shiny module for scan1 analysis and plots.

UI for scan1 analyses and summaries to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

