% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScan1Plot.R
\name{shinyScan1Plot}
\alias{shinyScan1Plot}
\alias{shinyScan1PlotOutput}
\alias{shinyScan1PlotUI}
\title{Shiny coefficient analysis and plot module}
\usage{
shinyScan1Plot(input, output, session, win_par, pmap_obj, phe_df, cov_mx,
  probs_obj, K_chr)

shinyScan1PlotUI(id)

shinyScan1PlotOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{win_par, phe_df, cov_mx, probs_obj, K_chr}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 coefficient plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

