% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScan1SNP.R
\name{shinyScan1SNP}
\alias{shinyScan1SNP}
\alias{shinyScan1SNPOutput}
\alias{shinyScan1SNPUI}
\title{Shiny scan1 SNP analysis and plot module}
\usage{
shinyScan1SNP(input, output, session, snp_par, chr_pos, pheno_names,
  snp_scan_obj, snp_action = shiny::reactive({     "basic" }))

shinyScan1SNPUI(id)

shinyScan1SNPOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{snp_par, chr_pos, pheno_names, snp_scan_obj, snp_action}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

