% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScatterPlot.R
\name{shinyScatterPlot}
\alias{shinyScatterPlot}
\title{Shiny scatter plot module}
\usage{
shinyScatterPlot(input, output, session, med_par, patterns, geno_max, peak_mar,
  med_ls, mediate_obj, phe_mx, cov_df, K_chr, allele_info)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{med_par, patterns, geno_max, peak_mar, med_ls, mediate_obj, phe_mx, cov_df, K_chr, allele_info}{reactive arguments}

\item{id}{shiny identifier}
}
\description{
Shiny module for scatter plots, with interfaces \code{shinyScatterPlotUI} and  \code{shinyScatterPlotOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
