% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySetup.R
\name{shinySetup}
\alias{shinySetup}
\alias{shinySetupInput}
\alias{shinySetupUI}
\title{Shiny setup for DOQTL selection}
\usage{
shinySetup(input, output, session, pheno_typer, peaks_tbl, pmap_obj,
  analyses_tbl, cov_df, pheno_data, projects_info)

shinySetupInput(id)

shinySetupUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{pheno_typer, peaks_tbl, pmap_obj, analyses_tbl, cov_df, pheno_data, projects_info}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1}}}
}
\description{
Shiny module for phenotype selection.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
