% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySetup.R
\name{shinySetup}
\alias{shinySetup}
\title{Shiny setup for DOQTL selection}
\usage{
shinySetup(
  input,
  output,
  session,
  pheno_typer,
  peaks_tbl,
  pmap_obj,
  analyses_tbl,
  cov_df,
  projects_info
)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{pheno_typer, peaks_tbl, pmap_obj, analyses_tbl, cov_df, projects_info}{reactive arguments}
}
\description{
Shiny module for phenotype selection, with interfaces \code{shinySetupInput} and  \code{shinySetupUI}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
