% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySNPFeature.R
\name{shinySNPFeature}
\alias{shinySNPFeature}
\title{Shiny SNP Features in SNP Region module}
\usage{
shinySNPFeature(
  id,
  snp_par,
  chr_pos,
  snp_scan_obj,
  snpinfo,
  top_snps_tbl,
  gene_exon_tbl,
  snp_action = shiny::reactive({
     "basic"
 })
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_par, chr_pos, snp_scan_obj, top_snps_tbl, snpinfo, gene_exon_tbl, snp_action}{reactive arguments}
}
\description{
Shiny module for scan1 analysis and plots, with interfaces \code{shinySNPFeatureInput}, \code{shinySNPFeatureUI} and  \code{shinySNPFeatureOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
