% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTopFeature.R
\name{shinyTopFeature}
\alias{shinyTopFeature}
\alias{shinyTopFeatureUI}
\title{Shiny Top Features in SNP Region module}
\usage{
shinyTopFeature(input, output, session, chr_pos, scan_obj, top_snps_tbl,
  gene_exon_tbl)

shinyTopFeatureUI(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{scan_obj, top_snps_tbl, gene_exon_tbl}{reactive arguments}

\item{id}{identifier for \code{\link{shinyScan1SNP}} use}
}
\description{
Shiny module for scan1 analysis and plots.

UI for top features to use in shiny module.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

