% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpSumServer.R
\name{snpSumServer}
\alias{snpSumServer}
\alias{snpSumInput}
\alias{snpSumUI}
\alias{snpSumOutput}
\title{Shiny SNP summary module}
\usage{
snpSumServer(id, snp_list, project_df)

snpSumInput(id)

snpSumUI(id)

snpSumOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{snp_list, project_df}{reactive arguments}
}
\value{
No return value; called for side effects.
}
\description{
Shiny module for SNP summary, with interfaces \code{shinySNPInput}, \code{shinySNPUI} and  \code{shinySNPOutput}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
