% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_pat_plot.R
\name{top_pat_plot}
\alias{top_pat_plot}
\title{Top Pattern Plot}
\usage{
top_pat_plot(pheno, scan_obj, chr_id, map, xlim, drop.hilit = 1.5,
  facet = "pheno", snp_action = "basic", cex = 2, ...)
}
\arguments{
\item{pheno}{name of phenotype for effect scan}

\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{xlim}{x limits for plot}

\item{facet}{facet by one of c("pheno","pattern")}

\item{snp_action}{character string for plot}

\item{cex}{size of plot symbols}

\item{fill.null}{null plot if no data}
}
\description{
Plot of top patterns.
}
\examples{
\dontrun{top_pat_plot(pheno, scan_obj, xlim)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
