% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{top_snp_asso}
\alias{top_snp_asso}
\title{Top SNP Association Plot Wrapper}
\usage{
top_snp_asso(pheno, scan_obj, xlim, snp_action = "basic")
}
\arguments{
\item{pheno}{name of phenotype for effect scan}

\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{xlim}{x limits for plot}

\item{snp_action}{character string for plot}
}
\description{
Top SNP Association Plot Wrapper
}
\examples{
\dontrun{top_snp_asso(pheno, scan_obj, xlim)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}

