% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_snp_asso.R
\name{top_snp_asso}
\alias{top_snp_asso}
\title{Top SNP Association Plot Wrapper}
\usage{
top_snp_asso(scan_obj, xlim, snp_action = "basic",
  phename = dimnames(scan_obj$lod)[[2]], drop.hilit = NULL)
}
\arguments{
\item{scan_obj}{object of class \code{\link[qtl2scan]{scan1}}}

\item{xlim}{x limits for plot}

\item{snp_action}{character string for plot}

\item{phename}{names of phenotypes}

\item{drop.hilit}{hilight SNPs within this value of max}

\item{pheno}{name of phenotype for effect scan}
}
\description{
Top SNP Association Plot Wrapper
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
